# openapi.yaml
openapi: "3.1.0"
info:
  title: MTS IaaS Cloud
  description: |
    Управление виртуальными машинами, проектами, пользователями и сетями
    на платформе MTS IaaS Cloud. Этот файл — единый источник истины (source of truth)
    для бэкенда и фронтенда.
  version: "1.0.0"
  contact:
    name: MTS IaaS Team
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: http://localhost:4010
    description: Prism mock server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

    # ── Ошибки ──────────────────────────────────────
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
      example:
        detail: "Resource not found"
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

    # ── Auth / User ─────────────────────────────────
...
    UserResponse:
      type: object
      required: [id, email, role, is_active]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        is_active:
          type: boolean
          example: true
...
    AdminUserInfo:
      allOf:
        - $ref: "#/components/schemas/UserResponse"
        - type: object
          required: [created_at]
          properties:
            created_at:
              type: string
              format: date-time

    # ── VM ──────────────────────────────────────────
...
    ProjectStatus:
      type: string
      enum:
        - CREATING
        - ACTIVE
        - INACTIVE
        - DELETING
        - ERROR
...
    ProjectResponse:
      type: object
      required: [id, name, cpu_quota, ram_quota, ssd_quota, cpu_used, ram_used, ssd_used, is_allocated, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_id:
          type: string
          format: uuid
          nullable: true
        cpu_quota:
          type: integer
        ram_quota:
          type: integer
        ssd_quota:
          type: integer
        cpu_used:
          type: integer
        ram_used:
          type: integer
        ssd_used:
          type: integer
        is_allocated:
          type: boolean
        status:
          $ref: "#/components/schemas/ProjectStatus"
      example:
        id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        name: "production-project"
        owner_id: "550e8400-e29b-41d4-a716-446655440000"
        cpu_quota: 8
        ram_quota: 16384
        ssd_quota: 100
        cpu_used: 2
        ram_used: 4096
        ssd_used: 50
        is_allocated: true
        status: "ACTIVE"
...
    AdminProjectCreateRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          example: "admin-project"
        cpu_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          default: 8
        ram_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          default: 16384
        ssd_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          default: 100
    AdminProjectResponse:
      allOf:
        - $ref: "#/components/schemas/ProjectResponse"
        - type: object
          required: [created_at]
          properties:
            status:
              type: string
            created_at:
              type: string
              format: date-time
    AdminVMCreateRequest:
      type: object
      required: [name, cpu, ram, ssd, project_id]
      properties:
        name:
          type: string
          minLength: 1
          example: "admin-vm-01"
        cpu:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 4
        ram:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 8192
        ssd:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 100
        project_id:
          type: string
          format: uuid
    AdminVMResponse:
      type: object
      required: [id, name, project_id, cpu, ram, ssd, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project_id:
          type: string
          format: uuid
        cpu:
          type: integer
        ram:
          type: integer
        ssd:
          type: integer
        status:
          type: string
        docker_container_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AdminVMInfoResponse:
      allOf:
        - $ref: "#/components/schemas/AdminVMResponse"
        - type: object
          properties:
            cpu_usage_percent:
              type: number
              format: float
              default: 0.0
            ram_usage_percent:
              type: number
              format: float
              default: 0.0
            ssd_usage_percent:
              type: number
              format: float
              default: 0.0
            network_in_bytes:
              type: integer
              default: 0
            network_out_bytes:
              type: integer
              default: 0
    AdminServersListResponse:
      type: object
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: "#/components/schemas/AdminVMResponse"
    AdminProjectsListResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/AdminProjectResponse"

    # ── Job ─────────────────────────────────────────
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Resource not found"
...
    BadRequest:
      description: Invalid input (malformed HTTP, body parse error, etc.)
      content:
        text/plain:
          schema:
            type: string
          example: "Invalid HTTP request received."
        text/plain; charset=utf-8:
          schema:
            type: string
          example: "Invalid HTTP request received."
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
paths:
...
  /project/admin/add/{user_id}:
    post:
      operationId: adminAddProject
      summary: Создать проект для пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProjectCreateRequest"
      responses:
        "201":
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /project/admin/disable/{project_id}:
    post:
      operationId: adminDisableProject
      summary: Отключить проект
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Проект отключён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /project/admin/active/{project_id}:
    post:
      operationId: adminActivateProject
      summary: Активировать проект
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Проект активирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /project/admin/delete/{project_id}:
    delete:
      operationId: adminDeleteProject
      summary: Удалить проект (admin)
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Проект удалён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /project/admin/info/{project_id}:
    get:
      operationId: adminGetProjectInfo
      summary: Информация о проекте
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/admin/info/{user_id}:
    get:
      operationId: adminListProjectsByUser
      summary: Проекты пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Список проектов пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectsListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/admin/disabled:
    get:
      operationId: adminListDisabledProjects
      summary: Отключённые проекты
      tags: [admin]
      responses:
        "200":
          description: Список отключённых проектов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectsListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ═══════════════════════════════════════════════════
  #  Admin — Servers (VM)
  # ═══════════════════════════════════════════════════
  /server/admin/add/{user_id}:
    post:
      operationId: adminAddServer
      summary: Создать VM для пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVMCreateRequest"
      responses:
        "201":
          description: VM создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVMResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /server/admin/disable/{server_id}:
    post:
      operationId: adminDisableServer
      summary: Отключить VM
      tags: [admin]
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: VM отключена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVMResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /server/admin/active/{server_id}:
    post:
      operationId: adminActivateServer
      summary: Активировать VM
      tags: [admin]
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: VM активирована
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVMResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /server/admin/delete/{server_id}:
    delete:
      operationId: adminDeleteServer
      summary: Удалить VM (admin)
      tags: [admin]
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: VM удалена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVMResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /server/admin/info/{server_id}:
    get:
      operationId: adminGetServerInfo
      summary: Информация о VM с нагрузкой
      tags: [admin]
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: VM info with load metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVMInfoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /servers/admin/disabled:
    get:
      operationId: adminListDisabledServers
      summary: Отключённые серверы
      tags: [admin]
      responses:
        "200":
          description: Список отключённых серверов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminServersListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ═══════════════════════════════════════════════════
  #  Admin — Users
  # ═══════════════════════════════════════════════════
  /user/admin/disable/{user_id}:
    post:
      operationId: adminDisableUser
      summary: Отключить пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Пользователь отключён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /user/admin/active/{user_id}:
    post:
      operationId: adminActivateUser
      summary: Активировать пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Пользователь активирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /user/admin/delete/{user_id}:
    delete:
      operationId: adminDeleteUser
      summary: Удалить пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Пользователь удалён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ═══════════════════════════════════════════════════
  #  Admin — Server lists
  # ═══════════════════════════════════════════════════
  /servers/admin/info/user/{user_id}:
    get:
      operationId: adminListServersByUser
      summary: VM пользователя
      tags: [admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Список VM пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminServersListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /servers/admin/info/project/{project_id}:
    get:
      operationId: adminListServersByProject
      summary: VM в проекте
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Список VM в проекте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminServersListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ═══════════════════════════════════════════════════
  #  Agent
  # ═══════════════════════════════════════════════════
...